refactor: extract order patching logic into utility function
